Removed CommandManager | Overhauls | Added signals | Added to Fc default config | Fcs delegating more to block managers
Overhauls:
- Command and Block ids are now bytes
  - Values set relative to their containers
    - Blocks now set their Commands' unique ids
- Block IsExecuting func turned into property
- Unique Block names handled by UniqueKeyGenerator
- Split from ICommandSource: IHasCommands and ICommandRemovable
- BlockLogicManager no longer an ICommandSource
  - It does expect you to give it an IBlockManager so it can properly track and execute things, though
- Switched out some linq statements for the sake of Unity 2022.3
- Removed unnecessary usings

+++ Assets/Hyphlow/Editor/Scripts/FlowchartWindow/FcWindowComponents/FcWindowRepaintTriggerer.cs
@@ -74,7 +74,7 @@ namespace AtMycelia.Hyphlow.EditorUtils.FcWindow
             TriggerRepaint();
         }
 
-        public void OnPostBlockDeletion(ushort blockId)
+        public void OnPostBlockDeletion(byte blockId)
         {
             TriggerRepaint();
         }
